fix: fix bugs where input not exist movie
