+++ src/Components/Pages/DetailPage/DetailPage.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { useParams } from "react-router-dom";
-import { movieAPI } from "../../../apiService/UserServices";
+import { movieAPI } from "../../../apiService/MovieService";
 import { appSlice } from "../../../store/reducers/AppSlice";
 import { useAppDispatch } from "../../../hooks/redux";
