+++ src/Components/UserCard/UserCard.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { IMovie } from "../../apiService/UserServices";
+import { IMovie } from "../../apiService/MovieService";
 import { Link } from "react-router-dom";
 import { useAppDispatch, useAppSelector } from "../../hooks/redux";
 import { appSlice } from "../../store/reducers/AppSlice";
