+++ src/Components/UserViewer/UserViewer.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { movieAPI } from "../../apiService/UserServices";
+import { movieAPI } from "../../apiService/MovieService";
 import UserCard from "../UserCard/UserCard";
 import { useAppSelector } from "../../hooks/redux";
 import LimitSetter from "../LimitSetter/LimitSetter";
@@ -7,7 +7,7 @@ import NewPagination from "../Pagination/Pagination";
 
 const UserViewer: FC = () => {
   const { searchQuery, limit, currentPage } = useAppSelector(
-    (state) => state.appReducer,
+    (state) => state.appReducer
   );
   const { data, isLoading } = movieAPI.useFetchAllMoviesQuery({
     limit: limit,
@@ -20,9 +20,9 @@ const UserViewer: FC = () => {
       style={{ display: "flex", alignItems: "center", flexDirection: "column" }}
     >
       {!isLoading && <LimitSetter />}
-      {!isLoading && <NewPagination />}
+      {!isLoading && <NewPagination items={data?.data.movie_count} />}
       {isLoading && <p>Loading...</p>}
-      {!isLoading && data?.data.movies.length === 0 && <p>Not users</p>}
+      {!isLoading && data?.data.movie_count === 0 && <p>Not exist movie...</p>}
       {data?.data.movies &&
         data.data.movies.map((movie) => {
           return (
