+++ src/store/reducers/AppSlice.type.ts
@@ -1,0 +1,6 @@
+export interface AppState {
+  limit: number;
+  isDetailOpen: boolean;
+  searchQuery: string;
+  currentPage: number;
+}
